complete first edition on backup algorithm

+++ fw_utils/clear_old_file.py
@@ -4,7 +4,8 @@ import re
 import time
 
 
-def clear_old_backup(scan_dir: str, pattern: str = '.*', age_days: int = 7, minimal_safe_files: int = 0):
+def clear_old_backup(scan_dir: str, pattern: str = '.*', age_days: int = 7, minimal_safe_files: int = 0,
+                     dont_delete_file: str = None):
     days_before = (time.time() - age_days * 86400)
     template = re.compile(pattern)
     files_lists = {}
@@ -15,7 +16,7 @@ def clear_old_backup(scan_dir: str, pattern: str = '.*', age_days: int = 7, mini
     logging.debug(f"list found files: {files_lists}")
     if len(files_lists) > minimal_safe_files:
         for stat_time, file_path in files_lists.items():
-            if stat_time < days_before:
+            if stat_time < days_before and file_path != dont_delete_file:
                 try:
                     os.remove(file_path)
                 except Exception as io:
